fix parse() throws error on valid ReflectOnly Blink extended attributes
