+++ lib/productions/extended-attributes.js
@@ -1,7 +1,40 @@
 import { Base } from "./base.js";
 import { ArrayBase } from "./array-base.js";
-import { list, identifiersOrStrings, argument_list, autoParenter } from "./helpers.js";
+import { Token } from "./token.js";
+import { list, argument_list, autoParenter } from "./helpers.js";
 import { validationError } from "../error.js";
+
+/**
+ * @param {import("../tokeniser").Tokeniser} tokeniser
+ */
+function identifiers(tokeniser) {
+  const ids = list(tokeniser, { parser: Token.parser(tokeniser, "identifier"), listName: "identifier list" });
+  if (!ids.length) {
+    tokeniser.error("Expected identifiers but none found");
+  }
+  return ids;
+}
+/**
+ * 
+ * @param {import("../tokeniser").Tokeniser} tokeniser 
+ * @param {string} tokenName 
+ */
+function tokens(tokeniser, tokenName) {
+  const toks = list(tokeniser, {
+    parser: Token.parser(tokeniser, tokenName),
+    listName: tokenName + " list"
+  });
+  if (!toks.length) {
+    tokeniser.error(`Expected ${tokenName}s but none found`);
+  }
+  return toks;
+}
+
+// This will allow a set of strings to be parsed
+ function identifiersOrStrings(tokeniser) {
+  return tokens(tokeniser, "identifier") || tokens(tokeniser, "string");
+}
+
 
 class ExtendedAttributeParameters extends Base {
   /**
