+++ lib/productions/helpers.js
@@ -1,6 +1,5 @@
 import { Type } from "./type.js";
 import { Argument } from "./argument.js";
-import { Token } from "./token.js";
 import { ExtendedAttributes, SimpleExtendedAttribute } from "./extended-attributes.js";
 import { Operation } from "./operation.js";
 import { Attribute } from "./attribute.js";
@@ -43,9 +42,7 @@ export function list(tokeniser, { parser, allowDangler, listName = "list" }) {
   return items;
 }
 
-/**
- * @param {import("../tokeniser").Tokeniser} tokeniser
- */
+
 export function const_value(tokeniser) {
   return tokeniser.consume("true", "false", "Infinity", "-Infinity", "NaN", "decimal", "integer");
 }
@@ -107,17 +104,6 @@ export function primitive_type(tokeniser) {
   if (base) {
     return new Type({ source, tokens: { base } });
   }
-}
-
-/**
- * @param {import("../tokeniser").Tokeniser} tokeniser
- */
-export function identifiers(tokeniser) {
-  const ids = list(tokeniser, { parser: Token.parser(tokeniser, "identifier"), listName: "identifier list" });
-  if (!ids.length) {
-    tokeniser.error("Expected identifiers but none found");
-  }
-  return ids;
 }
 
 /**
@@ -288,26 +274,4 @@ export function autoParenter(data, parent) {
       return true;
     }
   });
-}
-
-/**
- * Returns a proxy that auto-assign `parent` field.
- * @template T
- * @param {T} tokeniser
- * @param {*} [tokenName] either can be a string or identifier
- * @return {T}
- */
-function tokens(tokeniser, tokenName) {
-  const toks = list(tokeniser, {
-    parser: Token.parser(tokeniser, tokenName),
-    listName: tokenName + " list"
-  });
-  if (!toks.length) {
-    tokeniser.error(`Expected ${tokenName}s but none found`);
-  }
-  return toks;
-}
-// This will allow a set of strings to be parsed
-export function identifiersOrStrings(tokeniser) {
-  return tokens(tokeniser, "identifier") || tokens(tokeniser, "string");
 }
